requirements.txt
azure-functions
azure-storage-blob==12.21.0
